# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/resetPassword": {
      "put": {
        "operationId": "AuthController_setPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/auth/updateFCM": {
      "put": {
        "operationId": "AuthController_setFCM",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFCMTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "authentication"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/lifter": {
      "post": {
        "operationId": "LifterController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLifterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "lifter"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "LifterController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifter"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/lifter/{id}": {
      "get": {
        "operationId": "LifterController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifter"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "LifterController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLifterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifter"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "LifterController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "lifter"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin": {
      "post": {
        "operationId": "AdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AdminController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/notification/send-push-notification": {
      "post": {
        "operationId": "NotificationController_sendNotification",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "msg",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
...
  },
  "info": {
    "title": "BENTECH",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "kaur@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678",
            "description": "Minimum 8 characters password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "lifter",
            "enum": [
              "lifter",
              "client",
              "admin"
            ]
          },
          "name": {
            "type": "string",
            "example": "John Sharma"
          },
          "email": {
            "type": "string",
            "example": "john@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "12345678",
            "description": "Minimum 8 characters required"
          },
          "contactNo": {
            "type": "string",
            "example": "989-999-9292"
          },
          "deviceType": {
            "type": "string",
            "example": "android"
          },
          "notificationToken": {
            "type": "string",
            "example": "dfdsfdsffdsfdsfsd32432dbsdg7y432jfnduy329824y"
          }
        },
        "required": [
          "role",
          "name",
          "email",
          "password",
          "contactNo",
          "deviceType",
          "notificationToken"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "kaur@gmail.com"
          },
          "newPassword": {
            "type": "string",
            "example": "12345678",
            "description": "Minimum 8 characters password"
          }
        },
        "required": [
          "email",
          "newPassword"
        ]
      },
      "UpdateFCMTokenDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "kaur@gmail.com"
          },
          "token": {
            "type": "string",
            "example": "1234fetrt5678"
          }
        },
        "required": [
          "email",
          "token"
        ]
      },
...
      "CreateLifterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "email": {
            "type": "string",
            "example": "xyz@gmail.com"
          },
          "contactNo": {
            "type": "string",
            "example": "998-222-9494"
          },
          "address": {
            "type": "string",
            "example": "plot no-123, phase 8"
          },
          "vehicleNo": {
            "type": "string",
            "example": "24fsd4321266"
          },
          "Status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "contactNo",
          "address",
          "vehicleNo",
          "Status"
        ]
      },
      "UpdateLifterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "contactNo": {
            "type": "string",
            "example": "998-222-9494"
          },
          "address": {
            "type": "string",
            "example": "plot no-123, phase 8"
          },
          "vehicleNo": {
            "type": "string",
            "example": "24fsd4321266"
          },
          "Status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name",
          "contactNo",
          "address",
          "vehicleNo",
          "Status"
        ]
      },
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "email": {
            "type": "string",
            "example": "xyz@gmail.com"
          },
          "contactNo": {
            "type": "string",
            "example": "998-222-9494"
          },
          "address": {
            "type": "string",
            "example": "house no-123, sector 8"
          },
          "Status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "contactNo",
          "address",
          "Status"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "contactNo": {
            "type": "string",
            "example": "998-222-9494"
          },
          "address": {
            "type": "string",
            "example": "house no-123, sector 8"
          },
          "Status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        },
        "required": [
          "name",
          "contactNo",
          "address",
          "Status"
        ]
      },
...
    }
  }
}